Move replay and camera to plugins
